# docs/swagger.yaml
definitions:
...
  routes.AutentiaPersonParams:
    properties:
      birthdate:
        type: string
      country:
        type: string
      description:
        type: string
      dni:
        type: string
      gender:
        type: string
      id:
        type: string
      institution:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      name:
        type: string
      names:
        type: string
      nro_audit:
        type: string
      users_id:
        type: string
      version_change:
        type: string
    required:
    - dni
    type: object
...
  routes.MessageResponse:
    properties:
      code:
        type: integer
      details:
        type: string
    type: object
...
  routes.responsePerson:
    properties:
      data: {}
    type: object
...
  routes.responseStorage:
    properties:
      message:
        type: string
      result:
        $ref: '#/definitions/routes.storageResult'
      status:
        type: integer
    type: object
...
  routes.storageResult:
    properties:
      code:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  description: api autentia manager
  title: Autentia Manager Documentation
  version: "1"
paths:
...
  /autentia/persons/{dni}:
    get:
      consumes:
      - application/json
      description: autentia person dni get
      parameters:
      - description: dni
        in: path
        name: dni
        required: true
        type: string
      - description: country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person dni get
      tags:
      - autentia persons
  /autentia/persons/all/{dni}:
    get:
      consumes:
      - application/json
      description: autentia person get
      parameters:
      - description: dni
        in: path
        name: dni
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person get
      tags:
      - autentia persons
  /autentia/persons/compare/{dni}:
    get:
      consumes:
      - application/json
      description: autentia person compare get
      parameters:
      - description: dni
        in: path
        name: dni
        required: true
        type: string
      - description: version
        in: query
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person compare get
      tags:
      - autentia persons
  /autentia/persons/document:
    post:
      consumes:
      - multipart/form-data
      description: autentia person post document
      parameters:
      - in: formData
        name: delimiter
        type: string
      - in: formData
        name: descriptions
        type: string
      - in: formData
        name: dni
        type: string
      - in: formData
        name: name
        type: string
      - in: formData
        name: person_id
        type: string
      - in: formData
        name: uri
        type: string
      - description: people
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responseStorage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person post document
      tags:
      - autentia persons
  /autentia/persons/document/status:
    get:
      consumes:
      - application/json
      description: autentia person status get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person status get
      tags:
      - autentia persons
  /autentia/persons/document/version/{dni}:
    get:
      consumes:
      - application/json
      description: autentia person document version get
      parameters:
      - description: dni
        in: path
        name: dni
        required: true
        type: string
      - description: version
        in: query
        name: version
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person document version get
      tags:
      - autentia persons
  /autentia/persons/documents:
    delete:
      consumes:
      - application/json
      description: autentia person document delete
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person document delete
      tags:
      - autentia persons
  /autentia/persons/documents/{dni}:
    get:
      consumes:
      - application/json
      description: autentia person document get
      parameters:
      - description: dni
        in: path
        name: dni
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person document get
      tags:
      - autentia persons
  /autentia/persons/manager/new:
    post:
      consumes:
      - application/json
      description: autentia person post new
      parameters:
      - description: person
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/routes.AutentiaPersonParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person post new
      tags:
      - autentia persons
  /autentia/persons/manager/old:
    post:
      consumes:
      - application/json
      description: autentia person post old
      parameters:
      - description: person
        in: body
        name: person
        required: true
        schema:
          $ref: '#/definitions/routes.AutentiaPersonParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.responsePerson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/routes.MessageResponse'
      security:
      - BarerToken: []
      summary: autentia person post old
      tags:
      - autentia persons
...
securityDefinitions:
  BarerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"